wxGTK: Add private options to bring back traditional scrollbar(always visible).
GTK3 defaults to enable the 'overlay-scrolling' features, that is, only show scrollbar when the mouse/touching
hover the scrollbar area. It is undesirable to many of our clients.
Yet wxWidgets does NOT exposing any API for us to do the GTK customization.
So we need to add our 'CMExtension' to workaround this problem.

This MAY be a temporary solution. We may make a new patch in the future by introducing a new properly designed
API to wxWidgets for this sort of native/per-platform customization.